refactor(care-instructions): apply conventions
