--- src/app/care_instructions/components/CareInstructionButton.module.css
@@ -1,27 +1,0 @@
-/*
-CareInstructionButton.module.css
-Ojos Project
-*/
-.single_care_instruction {
-  width: 65dvw;
-  margin: 20px 0;
-  overflow-wrap: anywhere;
-  padding: 10px;
-  padding-right: 15px;
-  padding-left: 15px;
-  margin: 10px;
-  border-radius: 10px;
-  transition: transform 0.2s ease, box-shadow 0.2s ease;
-}
-
-.single_care_instruction:hover {
-  cursor: pointer;
-  transform: translateY(-10px);
-  box-shadow: 0 8px 15px rgba(111, 110, 110, 0.3);
-  
-}
-
-.tap_message {
-  font-size: 14px;
-  color: grey;
-}
