--- src/app/care_instructions/page.module.css
@@ -1,28 +1,0 @@
-/*
-/care_instructions/page.module.css
-Ojos Project
-*/
-.all_instructions_layout {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.back_button {
-  position: fixed;
-  top: 2%;
-  left: 2%;
-}
-
-.button_menu_container {
-  position: fixed;
-  bottom: 5%;
-  right: 5%;
-}
-
-.button_menu {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-end;
-}
